apply Rick patch
From email with JWW (20200509):

So, I had to change your script a bit to handle a case where one of my
fields on a line in the input file looked like this:

...,Hello World,...

In this case "Hello World" was transformed to "HelloWorld" in the
output.

I applied this very ugly patch to prevent that.

16,17c16,19
< 	} else if (match($0, /^[^ ,]+/)) {
< 	    printf("%s", substr($0, 1, RLENGTH))
---
> 	} else if (match($0, /^[^,]+/)) {
> 	    lvalue = substr($0, 1, RLENGTH)
> 	    sub(/ +$/, "", lvalue)
> 	    printf("%s", lvalue)

In the first line, I removed the space from the regex in the `match()`
call.  I could have stopped there but for the fact that now my match
string could have trailing spaces, so the other lines are supposed to
be handling that corner case.  (Notice, I used `sub` which requires an
lvalue as its 3rd argument.  That's why all the noise.  And I know I
could have used `gensub()` but that's not as portable (not POSIX, and
your program was already POSIX compliant[1]).)

Anyway, let know if you can find a neater way -- that is your puzzle
if you want it.  I'm pressing on with my ugly code. :)

Best, Rick

[1] https://pubs.opengroup.org/onlinepubs/009695399/utilities/awk.html